requirements.txt
Django>=4.2.3,<4.3
djangorestframework>=3.14.0,<3.15
python-dotenv>=1.0.1,<1.1
psycopg2>=2.9.9,<3
drf-spectacular>=0.26.5,<0.27
Pillow>=10.1.0,<10.2
django-filter